requirements.txt
fastapi
uvicorn
torch
torchvision
ultralytics
pillow
opencv-python
pyngrok